refactor(core): move inline tests from wasm.rs to dedicated test directory
